refactor: 머지 후 패키지 구조 분리 및 리팩터링
