--- be/src/main/java/com/issuetracker/dto/TestDto.java
@@ -1,13 +1,0 @@
-package com.issuetracker.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class TestDto {
-    private int a;
-    private String b;
-}
