+++ be/src/main/java/com/issuetracker/dto/issueList/FilterLabelDto.java
@@ -1,4 +1,4 @@
-package com.issuetracker.dto;
+package com.issuetracker.dto.issueList;
 
 import org.springframework.data.relational.core.mapping.Column;
 
@@ -9,12 +9,15 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class FilterMilestoneDto {
+public class FilterLabelDto {
     @Column("id")
-    private int milestoneId;
+    private int labelId;
     @Column("name")
-    private String milestoneName;
+    private String labelName;
+    @Column("background_color")
+    private String backgroundColor;
+    @Column("font_color")
+    private String fontColor;
     @Column("description")
     private String description;
 }
-
