+++ be/src/main/java/com/issuetracker/dto/issueList/FilterMilestoneDto.java
@@ -1,4 +1,4 @@
-package com.issuetracker.dto;
+package com.issuetracker.dto.issueList;
 
 import org.springframework.data.relational.core.mapping.Column;
 
@@ -9,15 +9,12 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class FilterLabelDto {
+public class FilterMilestoneDto {
     @Column("id")
-    private int labelId;
+    private int milestoneId;
     @Column("name")
-    private String labelName;
-    @Column("background_color")
-    private String backgroundColor;
-    @Column("font_color")
-    private String fontColor;
+    private String milestoneName;
     @Column("description")
     private String description;
 }
+
