+++ be/src/main/java/com/issuetracker/dto/issueList/IssueListDto.java
@@ -1,8 +1,7 @@
-package com.issuetracker.dto;
+package com.issuetracker.dto.issueList;
 
 import java.util.List;
 
-import com.issuetracker.dto.issue.IssueDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
